requirements.txt
# python 3.7.8
h5py==2.10.0
matplotlib==3.3.1
numpy==1.19.1
Pillow==7.2.0
tensorboard==2.3.0
torch==1.7.0
torchsummary==1.5.1
torchvision==0.8.1